ChatLink Support, Item Counts, Option Filters, oh my!
Added ChatLink support. Questlinks will inform you which Alts have the quest. Itemlinks will inform you where the item is dropped if it’s gear. If it’s a consumable or any other type of item you can use it will tell you how many you have and which Alts have them. In some cases, it will tell you if an item is a crafted item. “Item Source” can be toggled on/off in the Options panel.

Speaking of “…how many you have…”. Item counts! Yes, hover over a piece of Wool, a Bandage, or a healing potion etc. The tooltip will inform you which Alts have it, how much, and tell you if it’s in the Alt’s bank or bags. “Item Source, Item Count Per Character, and Total Item Count” can be toggled on/off in the Options panel.

Options… both “Already Known” and “Learnable By…” were at one point combined into one on/off switch. Now, they are separated and can be toggled separately. By default, “Include Already Known” is off and “Include Learnable By” is on. Makes for a cleaner tooltip as there is already a flood of info going to it.

All tooltips are now processed not just inside Altaholic but now outside and information is available from Player Bank and Bags, the Mailbox, Auction House, and Chat frames.